Migrate more setFoo(List) setters
Clean up more call sites, using BindingMap.ordered(). Also fixes
a few performance regressions introduced during original migration.

+++ openflowplugin-impl/src/main/java/org/opendaylight/openflowplugin/impl/protocol/serialization/actions/SetTpDstActionSerializer.java
@@ -19,6 +19,7 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.model.match.types.rev131026
 import org.opendaylight.yang.gen.v1.urn.opendaylight.model.match.types.rev131026.match.Icmpv6MatchBuilder;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.model.match.types.rev131026.match.layer._4.match.TcpMatchBuilder;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.model.match.types.rev131026.match.layer._4.match.UdpMatchBuilder;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 public class SetTpDstActionSerializer extends AbstractSetFieldActionSerializer {
     @Override
@@ -31,13 +32,13 @@ public class SetTpDstActionSerializer extends AbstractSetFieldActionSerializer {
             switch (proto) {
                 case ICMP: {
                     builder.setIcmpv4Match(new Icmpv4MatchBuilder()
-                            .setIcmpv4Code((short) (0xFF & port.getValue().toJava()))
+                            .setIcmpv4Code(Uint8.valueOf(0xFF & port.getValue().toJava()))
                             .build());
                     break;
                 }
                 case ICMPV6: {
                     builder.setIcmpv6Match(new Icmpv6MatchBuilder()
-                            .setIcmpv6Code((short) (0xFF & port.getValue().toJava()))
+                            .setIcmpv6Code(Uint8.valueOf(0xFF & port.getValue().toJava()))
                             .build());
                     break;
                 }
